fix: Scale years by opacity not color
In heingsight this is a much simpler approach, and even allows the removal
of the color library chroma-js from the project.  In all honestly this could
also facilitate the removal of recoil, as the global state is no longer
truly necessary, but I want to leave it with hope of more thoroughly using
recoil in the future.