fix: Simplify spring animations
The `from` and `to` states don't have to cover every back and forth,
only the original one. After that, we animate from the value of the
previous calculation of `to` to the current calculation of `to`.

---

react-spring is smart enough to animate from its current state to the
new `to` state.  The `from` state is only used for the very first
transition, when the component is first mounted.

Then, react-spring animates from the current state (which resulted from
the first calculation of the `to` state, to a new calculation of the
`to` state.